+++ MarkdownHelper/history.md
@@ -2,6 +2,7 @@
 
 ## v0.1.4
 
+* "save note to file button" now can handle relative links to files
 
 ## v0.1.3